fix(filenames): refactor filename generation for documents and deliverables
